refactor: extract logic to smaller components, fix TextInputQuestion call, and update questionnaire completion logic
Moved logic into smaller components. Fixed an error in calling the TextInputQuestion component. Updated the finishQuestionnaire logic to send a key-value object instead of a list.
