+++ .env
@@ -1,0 +1,1 @@
+VITE_URL_BACK = http://localhost:3000/
