+++ src/components/months-selector/MonthsSelector.jsx
@@ -51,9 +51,7 @@ const MonthsSelector = ({ onInput }) => {
     }
   }, []);
 
-
   return (
-
     <div className="outer-container" tabIndex={0}>
       <div className="container">
         <div className="column">
@@ -72,7 +70,6 @@ const MonthsSelector = ({ onInput }) => {
         </div>
       </div>
     </div>
-
   );
 };
