+++ src/pages/stepper/Stepper.jsx
@@ -8,80 +8,16 @@ import {
   FORCE_PLAN,
   NEXT,
 } from '../../utils/textConstant.js';
-import Steps from '../../components/steps/Steps.jsx';
-
-import MultiSteps from '../../components/steps/MultiSteps.jsx';
-
-//mover a otro archivo
-const MultiQuestion = ({ questions, setAnswer, setIsValidStep }) => {
-  const [currentAnswers, setCurrentAnswers] = useState({});
-  const [title, setTitle] = useState(null);
-  const [currentQuestions, setCurrentQuestions] = useState([]);
-
-  const handleInput = (currentQuestion, currentAnswer) => {
-    let updatedAnswers = currentAnswers;
-    if (!currentAnswer) {
-      delete updatedAnswers[currentQuestion.id];
-    } else {      
-      updatedAnswers = { ...updatedAnswers, [currentQuestion.id]: currentAnswer };      
-      setAnswer(currentQuestion, currentAnswer);
-    }
-    setCurrentAnswers(updatedAnswers);
-    const answersCount = Object.keys(updatedAnswers).length;
-    setIsValidStep(answersCount === currentQuestions.length);
-  };
-
-  useEffect(() => {
-    const titleQuestion = questions.find(q => q.id === 'title');
-    setTitle(titleQuestion ? titleQuestion.title : '');
-    setCurrentQuestions(questions.filter(q => q.id !== 'title' && q.id !== 'next'));
-  }, []);
-
-  return (
-    <>
-      <h3 className="sub-title">{title}</h3>
-      <div className="question-container">
-        {currentQuestions.map((question) => (
-          <div key={question.id}> 
-            <h3 className="sub-title">{question.question}</h3>
-            <MultiSteps
-              currentQuestion={question}
-              handleInput={handleInput}
-              currentAnswer={currentAnswers[question.id]}
-            />
-          </div>
-        ))}
-      </div>
-    </>
-  );
-};
-//mover a otro archivo
-const SingleQuestion = ({ question, setAnswer, setIsValidStep }) => {
-
-  const [currentAnswer, setCurrentAnswer] = useState(null);
-  const handleInput = (input) => {
-    if (!input) {      
-      setCurrentAnswer(input);
-      return setIsValidStep(false);
-    }    
-    setAnswer(question, input);
-    setCurrentAnswer(input);
-    setIsValidStep(true);    
-  };
-
-  return (
-    <div className="question-container">
-      <h3 className="sub-title">{question.question}</h3>
-      <Steps
-        currentQuestion={question}
-        handleInput={handleInput}
-        currentAnswer={currentAnswer}
-      />
-    </div>
-  );
-};
+import getRoutine from '../../services/questions/getRoutine.js';
+import { useNavigate } from 'react-router-dom';
+import { useAppProvider } from '../context-provider/AppProvider.jsx';
+import MultiQuestion from './components/multi-question/MultiQuestion.jsx';
+import SingleQuestion from './components/single-question/SingleQuestion.jsx';
 
 const Stepper = () => {
+  const { updateRoutine } = useAppProvider();
+
+  const navigate = useNavigate();
   //El nodo sería el objeto questions que tiene la o las preguntas actuales
   const [actualNode, setActualNode] = useState([]);
   
@@ -109,16 +45,13 @@ const Stepper = () => {
   const getNextNode = () => {
     const nextNode = getSimpleNextNode(); //obtengo el siguiente nodo común, el que no es parte de options
     const optionQuestion = actualNode.find(x=> x.input?.type === 'options'); //obtengo el objeto options si hay en este nodo 
-    
     if (optionQuestion) { //Si este nodo es uno de opciones   
       return manageOptionNode(nextNode,optionQuestion);
     }
     if (nextNode) {  // Checkeo si hay un next 
       return nextNode;
     }
-
     return checkNodeStack(); //Si no hay next veo si me quedó alguno en el stack pendiente  
-   
   };  
 
   const manageOptionNode = (nextNode,optionQuestion) =>{
@@ -127,7 +60,6 @@ const Stepper = () => {
     }
     //obtengo el next de la option elegida 
     var nextOption = getNextOptionNode(optionQuestion, answers[optionQuestion.id]);
-
     if (!nextOption) { //Si es null es porque hasta acá llegó el camino de la option y verifico si tengo para seguir otro camino guardado en el stack
       return checkNodeStack();
     }
@@ -143,11 +75,32 @@ const Stepper = () => {
     }
     return null;
   };
+
   const handleNext = () => {
     const nextNode = getNextNode();
-    setIsValidStep(false); 
-    return setActualNode(nextNode);
+    if (nextNode) { 
+      setIsValidStep(false); 
+      return setActualNode(nextNode);
+    }
+    finishQuestionnaire(answers);
   };
+
+  const finishQuestionnaire = async (answers) => {
+    console.log(answers);
+    try {
+      setLoading(true);
+      const newRoutine = await getRoutine(answers);
+      console.log(newRoutine);
+      updateRoutine(newRoutine);
+      await new Promise((resolve) => setTimeout(resolve, 500));
+    } catch (e) {
+      setError(e);
+    } finally {
+      navigate('/routine');
+      setLoading(false);
+    }
+  };
+
   const fetchInitialData = async () => {
     setLoading(true);
     setError(null);
@@ -170,7 +123,6 @@ const Stepper = () => {
     fetchInitialData();
   }, []);
 
-
   if (loading) {
     return (
       <div className="stepper-container">
@@ -186,7 +138,8 @@ const Stepper = () => {
       </div>
     );
   }
-  const getActualQuestion = () =>{
+  const getActualQuestion = () => {
+    
     return actualNode.find(x => x.id !== 'next');
   };
   return (
@@ -196,7 +149,7 @@ const Stepper = () => {
         {
           actualNode ? (
             actualNode.length > 2 ? (
-              <MultiQuestion 
+              <MultiQuestion
                 questions={actualNode} 
                 setAnswer={setNewAnswer} 
                 setIsValidStep = {setIsValidStep} 
@@ -220,4 +173,5 @@ const Stepper = () => {
   
   );
 };
+
 export default Stepper;
