+++ src/services/questions/getRoutine.js
@@ -1,6 +1,8 @@
-export default async function getRoutine (answers) {
+async function getRoutine (answers) {
+  console.log(answers);
   try {
     const answersBody = JSON.stringify(answers);
+    console.log(answersBody);
     const response = await fetch('http://localhost:3000/routine', {
       method: 'POST',
       headers: {
@@ -11,7 +13,6 @@ export default async function getRoutine (answers) {
     if (!response.ok) {
       throw new Error('Network response was not ok');
     }
-    
     const data = await response.json(); 
     return data; 
   } catch (err) {
@@ -19,3 +20,5 @@ export default async function getRoutine (answers) {
     throw err; 
   }
 }
+
+export default getRoutine;
